Power plant heat rate data (Ipython and Python)
The California state energy almanac contains power plant level heat rate
data. However, it is not easily accessed (only a table on the website).
This code scrapes the website for the data and produces a series of data
files by year. At this time I don't know how to form one dataset
(instead of one per year) but this improvement should be easy enough to
make when more time is available.